requirements.txt
# Core dependencies - Python 3.11 Compatible with nfl-data-py
pandas==1.5.3           # Required by nfl-data-py (<2.0)
requests==2.32.4        # Latest for API calls  
duckdb==1.3.2          # Latest for analytics database
nfl-data-py==0.3.3     # NFL data source
pyyaml==6.0.2          # Configuration files
python-dotenv==1.0.1   # Environment variables
click==8.1.7           # CLI framework
rich==13.9.4           # Terminal formatting
numpy==1.24.4          # Compatible with pandas 1.5.3
scipy==1.11.4          # Compatible with numpy 1.24.4

# Development dependencies
pytest==8.3.4          # Testing framework
pytest-cov==6.0.0      # Test coverage
ruff==0.9.1            # Fast linting
sqlfluff==3.3.1        # SQL formatting
ipython==8.18.1        # Interactive shell